+++ package/lib/lib/languages-list.ts
@@ -1,0 +1,16 @@
+export interface LanguageOption {
+    id: string;
+    label: string;
+}
+export const languagesList: Array<LanguageOption> = [
+    { id: 'javascript', label: 'JavaScript' },
+    { id: 'typescript', label: 'TypeScript' },
+    { id: 'python', label: 'Python' },
+    { id: 'java', label: 'Java' },
+    { id: 'csharp', label: 'C#' },
+    { id: 'ruby', label: 'Ruby' },
+    { id: 'go', label: 'Go' },
+    { id: 'php', label: 'PHP' },
+    { id: 'html', label: 'HTML' },
+    { id: 'css', label: 'CSS' },
+];