+++ package/vite.config.ts
@@ -46,6 +46,7 @@ export default defineConfig({
         '@tiptap/static-renderer',
         '@floating-ui/react',
         '@tiptap/extension-mention',
+        '@tiptap/extension-suggestion',
         '@radix-ui/react-dropdown-menu',
         '@radix-ui/react-popover',
         'lodash.isequal',
@@ -77,8 +78,9 @@ export default defineConfig({
           '@tiptap/extension-text-align': 'TiptapExtensionTextAlign',
           '@tiptap/extension-typography': 'TiptapExtensionTypography',
           '@tiptap/static-renderer': 'TiptapStaticRenderer',
-          '@tiptap/extension-mention':'TiptapExtensionMention'
-          
+          '@tiptap/extension-mention':'TiptapExtensionMention',
+          '@tiptap/extension-suggestion':'TiptapExtensionSuggestion'
+
         },
       },
     },