+++ demo/CHANGELOG.md
@@ -1,9 +1,19 @@
 # nexo-editor-demo
 
+## 0.0.3
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+- Updated dependencies
+  - nexo-editor@2.2.0
+
 ## 0.0.2
 
 ### Patch Changes
 
 - Updated dependencies
 - Updated dependencies
-  - nexo-editor@2.0.0
+  - nexo-editor@2.1.0